Revert "[Fix] Attempt at package extraction"
This reverts commit af88c4a5a86a385bf6d901703c13887a9411c30d.
