--- packages/connectors_shared/src/FontConnector.Shared.external.ts
@@ -1,20 +1,0 @@
-import { Id } from "./Connector.Shared.external";
-
-export type FontConnectorCapabilities = {
-  query: boolean;
-  detail: boolean;
-  preview: boolean;
-  filtering: boolean;
-};
-
-export enum FontPreviewFormat {
-  Square = "square",
-  Line = "line",
-}
-
-export interface FontFamily {
-  id: Id;
-  name: string;
-  fontStylesCount: number;
-  extensions: string[];
-}
