--- packages/connectors_shared/src/index.ts
@@ -1,3 +1,0 @@
-export * as Connector from "./Connector.Shared.external";
-export * as Font from "./FontConnector.Shared.external";
-export * as Media from "./MediaConnector.Shared.external";
