--- packages/connectors_shared/tsconfig.json
@@ -1,9 +1,0 @@
-{
-  "compilerOptions": {
-    "declaration": true,
-    "emitDeclarationOnly": true,
-    "outDir": "dist"
-  },
-  "include": ["src"],
-  "exclude": ["node_modules", "**/__tests__/*"]
-}
